In 10.3, the last sentence is true
Where it says "Consider the following sentences:" in the 3x4
dot world, the last statement (#10) is true. It says that
there exists a blue dot that is above a green dot. This is true
for the blue dot in the fourth column of the first row, because
it is above the green dot in the third column of the second row.
(Ignoring which columns they are in, similar to how left-of
ignores which rows the dots are in.)