Chapter 19 and 20 errata
Chapter 19 Exercises

Grammatical fix:
Exercise 6: Write 11160 as product of primes.
Should be: Write 11160 as the product of primes.
(Or "a product of primes".)

Exercise 19: Suppose that a | n and a | n
Should be: Suppose that a | n and a | m

Chapter 20

In inclusion-exclusion principle, exponent on -1 is off by one.
So needs to be either |I|-1 or |I|+1.

Was:
How many are there IS the first digit
Should be:
How many are there IF the first digit

Added missing parenthesis at the end of Hint.

For consistency, added hyphen between "inclusion-exclusion", as it was in the
section heading (and as it appears in other sources).

Took 0's out of math mode, because that causes the apostrophe to render as a prime,
and the s as a variable. (Not that you really need apostrophes here, but I won't
quibble that style bit.)

It's k 0s, not k 1s.